refactor(models): unify frontmatter field extraction in FrontmatterParser
The extractSummary(from content:) helper was duplicated across five
sites (TodayViewModel.swift L378 + L955, ArchiveView.swift L186 + L415,
TimelineService.swift L230, WeeklyRecapService.swift L121). All four
implementations were behaviourally identical: scan lines, match
"summary:" prefix, strip surrounding whitespace + matching quotes,
return nil on empty/missing.

The duplication was a latent bug source — any future field tweak (e.g.
adding summary_zh: for localised compilations) had to land in five
places at once. Adding a new field meant a 5x rename + retest.

Fix:
- Add DayPage/Models/FrontmatterParser.swift with a single static
  extractField(_ key:, from content:) that preserves the exact
  behaviour of all five removed copies.
- Replace all five call sites with FrontmatterParser.extractField(
  "summary", from: content).
- Delete the now-unused private extractSummary helpers in TodayViewModel,
  ArchiveView, TimelineService, and the static WeeklyRecapService.
  extractSummary (no external callers — grep confirmed).
- Register the new file in DayPage.xcodeproj/project.pbxproj (PBXBuildFile
  + PBXFileReference + Models group + Sources phase).

Behavioural equivalence: each removed helper used the exact same
algorithm. extractField("summary", ...) lands identical output on
existing daily page files; no user-visible change.

Verification: xcodebuild on iPhone 17 simulator — green.
